requirements.txt
# 基础
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# 文档加载
pdfplumber>=0.10.3
python-docx>=0.8.11
beautifulsoup4>=4.12.2

# 文本分块
langchain>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-community>=0.0.10

# 向量库 & 检索
chromadb>=0.4.22
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# LLM & 智谱
zhipuai>=2.1.0
openai>=1.0.0

# ======================
# Agent + MCP 新增依赖
# ======================
langchain-experimental>=0.1.0
tenacity>=8.2.0
psutil>=5.9.0
python-multipart>=0.0.6
